+++ server-actions-for-next-pages/src/browser.ts
@@ -4,6 +4,7 @@ import { EventSourceParserStream } from 'eventsource-parser/stream';
 type NextRpcCall = (...params: any[]) => any;
 
 let nextId = 1;
+
 async function* yieldServerSentEvents(response: Response, superjson: any) {
   if (!response.body) {
     throw new Error('Response body is null');
@@ -34,7 +35,53 @@ async function* yieldServerSentEvents(response: Response, superjson: any) {
   }
 }
 
-export function createRpcFetcher(url: string, method: string): NextRpcCall {
+export function createRpcFetcher(
+  url: string,
+  method: string,
+  isGenerator?: boolean,
+): NextRpcCall {
+  if (isGenerator) {
+    return async function* rpcGeneratorFetch(...args) {
+      const superjson = await import('superjson');
+      const { json, meta } = superjson.serialize(args);
+      const res = await fetch(url, {
+        method: 'POST',
+        body: JSON.stringify(
+          {
+            jsonrpc: '2.0',
+            id: nextId++,
+            method,
+            params: json as any[],
+            meta,
+          } satisfies JsonRpcRequest,
+          null,
+          2,
+        ),
+        headers: {
+          'content-type': 'application/json',
+        },
+      });
+
+      if (res.status === 502) {
+        const statusError = new Error('Unexpected HTTP status ' + res.status);
+        const json = await res.json();
+
+        if (json?.error && typeof json.error.message === 'string') {
+          let err = new Error(json.error.message);
+          Object.assign(err, json.error.data || {});
+          throw err;
+        }
+
+        throw statusError;
+      }
+
+      const contentType = res.headers.get('content-type');
+      if (contentType?.includes('text/event-stream')) {
+        yield* yieldServerSentEvents(res, superjson.default);
+      }
+    };
+  }
+
   return async function rpcFetch(...args) {
     const superjson = await import('superjson');
     const { json, meta } = superjson.serialize(args);